chore: setup shared primitives and move accounts/contracts there under test utils
