+++ .gitignore
@@ -1,4 +1,3 @@
 target/
 .idea/
-integration_logs/
 .DS_Store
