+++ crates/client/test-utils/src/lib.rs
@@ -3,23 +3,24 @@
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
-mod accounts;
-pub use accounts::Account;
+// Re-export from base-primitives for backwards compatibility
+pub use base_primitives::{
+    AccessListContract, Account, ContractFactory, DEVNET_CHAIN_ID, DoubleCounter, Logic, Logic2,
+    Minimal7702Account, MockERC20, Proxy, SimpleStorage, TransparentUpgradeableProxy,
+    build_test_genesis,
+};
 
 mod constants;
 pub use constants::{
-    BASE_CHAIN_ID, BLOCK_BUILD_DELAY_MS, BLOCK_TIME_SECONDS, DEFAULT_JWT_SECRET, GAS_LIMIT,
+    BLOCK_BUILD_DELAY_MS, BLOCK_TIME_SECONDS, DEFAULT_JWT_SECRET, GAS_LIMIT,
     L1_BLOCK_INFO_DEPOSIT_TX, L1_BLOCK_INFO_DEPOSIT_TX_HASH, NODE_STARTUP_DELAY_MS, NamedChain,
 };
-
-mod contracts;
-pub use contracts::{DoubleCounter, Minimal7702Account, MockERC20, TransparentUpgradeableProxy};
 
 mod engine;
 pub use engine::{EngineAddress, EngineApi, EngineProtocol, HttpEngine, IpcEngine};
 
 mod fixtures;
-pub use fixtures::{build_test_genesis, create_provider_factory, load_chain_spec};
+pub use fixtures::{create_provider_factory, load_chain_spec};
 
 mod flashblocks_harness;
 pub use flashblocks_harness::FlashblocksHarness;
