+++ crates/shared/access-lists/Cargo.toml
@@ -23,12 +23,12 @@ serde.workspace = true
 op-revm.workspace = true
 eyre.workspace = true
 reth-optimism-evm.workspace = true
+reth-optimism-chainspec.workspace = true
 alloy-consensus.workspace = true
 alloy-contract.workspace = true
-alloy-sol-macro = { workspace = true, features = ["json"] }
 alloy-sol-types.workspace = true
 reth-evm.workspace = true
-base-test-utils.workspace = true
+base-primitives = { workspace = true, features = ["test-utils"] }
 
 [[test]]
 name = "builder"
