+++ crates/shared/access-lists/tests/builder/main.rs
@@ -1,17 +1,19 @@
 //! Tests for ensuring the access list is built properly
 
-use std::collections::HashMap;
+use std::{collections::HashMap, sync::Arc};
 
 use alloy_consensus::Header;
 pub use alloy_primitives::{Address, B256, TxKind, U256};
-use alloy_sol_macro::sol;
 pub use alloy_sol_types::SolCall;
 use base_access_lists::FBALBuilderDb;
 pub use base_access_lists::FlashblockAccessList;
-use base_test_utils::{BASE_CHAIN_ID, load_chain_spec};
+use base_primitives::{
+    AccessListContract, ContractFactory, DEVNET_CHAIN_ID, SimpleStorage, build_test_genesis,
+};
 pub use eyre::Result;
 pub use op_revm::OpTransaction;
 use reth_evm::{ConfigureEvm, Evm};
+use reth_optimism_chainspec::OpChainSpec;
 use reth_optimism_evm::OpEvmConfig;
 use revm::{DatabaseCommit, context::result::ResultAndState, database::InMemoryDB};
 pub use revm::{
@@ -26,62 +28,10 @@ mod deployment;
 mod storage;
 mod transfers;
 
-sol!(
-    #[sol(rpc)]
-    AccessListContract,
-    concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/../../client/test-utils/contracts/out/AccessList.sol/AccessList.json"
-    )
-);
-
-sol!(
-    #[sol(rpc)]
-    ContractFactory,
-    concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/../../client/test-utils/contracts/out/ContractFactory.sol/ContractFactory.json"
-    )
-);
-
-sol!(
-    #[sol(rpc)]
-    SimpleStorage,
-    concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/../../client/test-utils/contracts/out/ContractFactory.sol/SimpleStorage.json"
-    )
-);
-
-sol!(
-    #[sol(rpc)]
-    Proxy,
-    concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/../../client/test-utils/contracts/out/Proxy.sol/Proxy.json"
-    )
-);
-
-sol!(
-    #[sol(rpc)]
-    Logic,
-    concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/../../client/test-utils/contracts/out/Proxy.sol/Logic.json"
-    )
-);
-
-sol!(
-    #[sol(rpc)]
-    Logic2,
-    concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/../../client/test-utils/contracts/out/Proxy.sol/Logic2.json"
-    )
-);
-
-/// Chain ID for Base Sepolia (re-export from test-utils for convenience)
-pub const BASE_SEPOLIA_CHAIN_ID: u64 = BASE_CHAIN_ID;
+/// Loads the test chain spec from the genesis configuration.
+fn load_chain_spec() -> Arc<OpChainSpec> {
+    Arc::new(OpChainSpec::from_genesis(build_test_genesis()))
+}
 
 /// Executes a list of transactions and builds a FlashblockAccessList tracking all
 /// account and storage changes across all transactions.
