+++ crates/shared/primitives/Cargo.toml
@@ -1,0 +1,46 @@
+[package]
+name = "base-primitives"
+description = "Shared primitives and test utilities for node-reth"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[features]
+default = []
+test-utils = [
+	"dep:alloy-consensus",
+	"dep:alloy-contract",
+	"dep:alloy-eips",
+	"dep:alloy-genesis",
+	"dep:alloy-signer",
+	"dep:alloy-signer-local",
+	"dep:alloy-sol-macro",
+	"dep:alloy-sol-types",
+	"dep:eyre",
+	"dep:op-alloy-network",
+	"dep:op-alloy-rpc-types",
+	"dep:serde_json",
+]
+
+[dependencies]
+alloy-primitives = { workspace = true, features = ["serde"] }
+
+# test-utils (optional)
+alloy-consensus = { workspace = true, features = ["std"], optional = true }
+alloy-contract = { workspace = true, optional = true }
+alloy-eips = { workspace = true, optional = true }
+alloy-genesis = { workspace = true, optional = true }
+alloy-signer = { workspace = true, optional = true }
+alloy-signer-local = { workspace = true, optional = true }
+alloy-sol-macro = { workspace = true, features = ["json"], optional = true }
+alloy-sol-types = { workspace = true, optional = true }
+eyre = { workspace = true, optional = true }
+op-alloy-network = { workspace = true, optional = true }
+op-alloy-rpc-types = { workspace = true, optional = true }
+serde_json = { workspace = true, optional = true }
