+++ crates/shared/primitives/contracts/.gitignore
