+++ crates/shared/primitives/contracts/script/DeployERC20.s.sol
