+++ crates/shared/primitives/contracts/src/DoubleCounter.sol
