+++ crates/shared/primitives/src/test_utils/contracts.rs
@@ -2,11 +2,6 @@
 //!
 //! This module provides pre-compiled contract bindings that can be used
 //! across different test crates without needing relative path references.
-//!
-//! Contract sources:
-//! - `DoubleCounter`: Custom test contract (src/DoubleCounter.sol)
-//! - `MockERC20`: Solmate's MockERC20 (lib/solmate)
-//! - `TransparentUpgradeableProxy`: OpenZeppelin's TransparentUpgradeableProxy (lib/openzeppelin-contracts)
 
 use alloy_sol_macro::sol;
 
@@ -38,4 +33,40 @@ sol!(
         env!("CARGO_MANIFEST_DIR"),
         "/contracts/out/Minimal7702Account.sol/Minimal7702Account.json"
     )
+);
+
+sol!(
+    #[sol(rpc)]
+    AccessListContract,
+    concat!(env!("CARGO_MANIFEST_DIR"), "/contracts/out/AccessList.sol/AccessList.json")
+);
+
+sol!(
+    #[sol(rpc)]
+    ContractFactory,
+    concat!(env!("CARGO_MANIFEST_DIR"), "/contracts/out/ContractFactory.sol/ContractFactory.json")
+);
+
+sol!(
+    #[sol(rpc)]
+    SimpleStorage,
+    concat!(env!("CARGO_MANIFEST_DIR"), "/contracts/out/ContractFactory.sol/SimpleStorage.json")
+);
+
+sol!(
+    #[sol(rpc)]
+    Proxy,
+    concat!(env!("CARGO_MANIFEST_DIR"), "/contracts/out/Proxy.sol/Proxy.json")
+);
+
+sol!(
+    #[sol(rpc)]
+    Logic,
+    concat!(env!("CARGO_MANIFEST_DIR"), "/contracts/out/Proxy.sol/Logic.json")
+);
+
+sol!(
+    #[sol(rpc)]
+    Logic2,
+    concat!(env!("CARGO_MANIFEST_DIR"), "/contracts/out/Proxy.sol/Logic2.json")
 );
