+++ crates/shared/primitives/src/test_utils/mod.rs
@@ -1,0 +1,13 @@
+//! Test utilities including accounts, genesis configuration, and contract bindings.
+
+mod accounts;
+pub use accounts::Account;
+
+mod genesis;
+pub use genesis::{DEVNET_CHAIN_ID, build_test_genesis};
+
+mod contracts;
+pub use contracts::{
+    AccessListContract, ContractFactory, DoubleCounter, Logic, Logic2, Minimal7702Account,
+    MockERC20, Proxy, SimpleStorage, TransparentUpgradeableProxy,
+};
