+++ lychee.toml
@@ -3,7 +3,7 @@ no_progress = false
 exclude_all_private = false
 accept = [200, 403] # 403 status code is often returned by private repos instead of 404
 exclude_path = [
-    "client/test-utils/contracts/lib"
+    "shared/primitives/contracts/lib"
 ]
 exclude = [
     'foo.bar',
